requirements.txt
gradio==5.12.0
langchain==0.3.15
langchain_community==0.3.15
langchain_openai==0.3.1
openai==1.60.0
pymupdf4llm==0.0.17
python-dotenv==1.0.1
